+++ tests/test_metadata.py
@@ -439,6 +439,7 @@ expected = {
     ],
     "json_files": [],
     "yaml_files": [
+        ".github/actions/pr-metadata/action.yml",
         ".github/funding.yml",
         ".github/workflows/autofix.yaml",
         ".github/workflows/autolock.yaml",