+++ changelog.md
@@ -5,6 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Replace workflow shell scripts with CLI commands.
+
 ## [5.5.0 (2026-01-29)](https://github.com/kdeldycke/workflows/compare/v5.4.0...v5.5.0)
 
 - Add `gha-utils update-exclude-newer` command to update the exclude-newer date in pyproject.toml.